new ft_memlen, insert now returns an int, fixed some leaks after insert, finished build_env

+++ mini/srcs/robin_hash/hash_functions/fnv1a.c
@@ -6,7 +6,7 @@
 /*   By: fldumas- <fldumas-@student.42angouleme.fr  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/05 16:30:04 by fldumas-          #+#    #+#             */
-/*   Updated: 2026/06/11 17:07:23 by fldumas-         ###   ########.fr       */
+/*   Updated: 2026/06/26 12:33:58 by fldumas-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,15 @@
 
 size_t	fnv1a(const void *key)
 {
-	size_t			hash;
+	uint64_t		hash;
 	unsigned char	*str;
 
 	hash = 14695981039346656037ULL;
 	str = (unsigned char *)key;
 	while (*str)
 	{
-		hash ^= (size_t)(*str++);
+		hash ^= (uint64_t)(*str++);
 		hash *= 1099511628211ULL;
 	}
-	return (hash);
+	return ((size_t)hash);
 }
